stopping augmentation work to get the planning baselines running
we want to see if there's really a transfer gap from floating boxes
part of that is ensuring that non floating boxes classifiers still
perform well in in planning within their traning envs, which might not
be true because we've changed a lot of small things. So first step is to
rerun those.