fix where we compute the relations
previously we were computing the relations from t-1 to predict the
velocity at time t, which produces the position at t+1. Instead now we
use the position at t and compute relations. So that manes the velocity
at t+1 is a function of the position and velocity at time t only.